refactor: share helper between getActiveStyle and getActiveListItem
Both selectors had the same shape: gather selected text blocks, return
the value of a given block-level property if every block agrees on it,
otherwise `undefined`. Extracted to a private
`getActiveBlockProperty(snapshot, key)` helper keyed on the property
name. The public `getActiveStyle` and `getActiveListItem` selectors
stay as named exports and forward to the helper.
