+++ packages/editor/src/selectors/selector.get-active-style.ts
@@ -1,35 +1,10 @@
 import type {PortableTextTextBlock} from '@portabletext/schema'
 import type {EditorSelector} from '../editor/editor-selector'
-import {getSelectedTextBlocks} from './selector.get-selected-text-blocks'
+import {getActiveBlockProperty} from './selector.get-active-block-property'
 
 /**
  * @public
  */
 export const getActiveStyle: EditorSelector<PortableTextTextBlock['style']> = (
   snapshot,
-) => {
-  if (!snapshot.context.selection) {
-    return undefined
-  }
-
-  const selectedTextBlocks = getSelectedTextBlocks(snapshot).map(
-    (block) => block.node,
-  )
-  const firstTextBlock = selectedTextBlocks.at(0)
-
-  if (!firstTextBlock) {
-    return undefined
-  }
-
-  const firstStyle = firstTextBlock.style
-
-  if (!firstStyle) {
-    return undefined
-  }
-
-  if (selectedTextBlocks.every((block) => block.style === firstStyle)) {
-    return firstStyle
-  }
-
-  return undefined
-}
+) => getActiveBlockProperty(snapshot, 'style')
